Sticky nav, img in projects, and created footer
Commit 3:
Created a sticky nav bar at the top of my website. Changed the background colour for the sticky header.
About section - added text align justify.
Projects - input images for all projects, gave the images an opcaity, created a headline for projects.
Contact -moved the linktree link to my contact section in my footer. Added icon for phone and icon for email. Created links so they open straight to call or open the email. Added a copyright. added the same background colour for the header as the footer.