Refactored CSS, added in <main>, added pseudo elements.
Added in the <main> and deleted a bunch of unnecessary styling - cleaning up the CSS.
Rearranged footer and adding in a google font icon to upgrade the 'CONTACT' icons.
Added before pseudo element to IDs that the nav links which added a gap at the top of the page and is not hidden behind the sticky nav. Changed the color of the links when hovered over using a pseudo selector.